Initial commit.
A snapshot of my doorbell notifier project. It's still WIP and thus quite messy. Sends the voltage read across a Pi Pico's ADC to an ESP8266 via UART using AT Messages. The ESP8266 then sends that data via TCP to the predefined TCP Server that is also included. When the server detects a voltage larger than a certain threshold, it plays a sound.